refactor c++ code: move more functions to header files, change file naming
