+++ inst/include/mapped_csr.hpp
@@ -1,6 +1,10 @@
+#ifndef MAPPED_CSR_HPP
+#define MAPPED_CSR_HPP
+
 #include <cstdint>
 #include <stddef.h>
 #include <armadillo>
+
 template< typename T>
 class MappedCSR {
 public:
@@ -8,16 +12,16 @@ public:
   MappedCSR(arma::uword n_rows,
             arma::uword n_cols,
             size_t nnz,
-            arma::uword * col_indices,
-            arma::uword * row_ptrs,
-            T * values):
+            arma::uword* col_indices,
+            arma::uword* row_ptrs,
+            T* values):
   n_rows(n_rows), n_cols(n_cols), nnz(nnz), col_indices(col_indices), row_ptrs(row_ptrs), values(values) {};
   const arma::uword n_rows;
   const arma::uword n_cols;
   const size_t nnz;
-  arma::uword * col_indices;
-  arma::uword * row_ptrs;
-  T * values;
+  arma::uword* col_indices;
+  arma::uword* row_ptrs;
+  T* values;
   std::pair<arma::uvec, arma::Col<T>> get_row(const arma::uword i) const {
     const arma::uword p1 = this->row_ptrs[i];
     const arma::uword p2 = this->row_ptrs[i + 1];
@@ -30,24 +34,4 @@ public:
 using dMappedCSR = MappedCSR<double>;
 using fMappedCSR = MappedCSR<float>;
 
-template< typename T>
-class MappedCSC {
-public:
-  MappedCSC();
-  MappedCSC(arma::uword n_rows,
-            arma::uword n_cols,
-            size_t nnz,
-            arma::uword * row_indices,
-            arma::uword * col_ptrs,
-            T * values):
-    n_rows(n_rows), n_cols(n_cols), nnz(nnz), row_indices(row_indices), col_ptrs(col_ptrs), values(values) {};
-  const arma::uword n_rows;
-  const arma::uword n_cols;
-  const size_t nnz;
-  arma::uword * row_indices;
-  arma::uword * col_ptrs;
-  T * values;
-};
-
-using dMappedCSC = MappedCSC<double>;
-using fMappedCSC = MappedCSC<float>;
+#endif /* MAPPED_CSR_HPP */
