+++ src/matrix_top_product.cpp
